feat: add crate::run; library is fully functional
