+++ .dockerignore
@@ -1,1 +1,1 @@
-utils/**/target
+target
