+++ TestFiles/to_subtraction.sam
@@ -1,0 +1,8 @@
+# This line should be discarded.
+# 1. Should not be removed because subtraction score is lower.
+# 2. Should be removed because subtraction score is higher.
+# 3. Should be removed because subtraction score is equal.
+@THIS line should also be discarded.
+HWI-ST1410:82:C2VAGACXX:7:1101:20066:1892	16	chrom_1	2346	255	100M1S	*	0	0	ACAAAGATTGCAGAAGGCGCGAAGAGGAGTGTTTGAGGCTCGCTACAGCTGAGGAAAAGGTTATGACAGTCGAAAGCGTGCTTAAGAAAGAGGCTGCGAAN	CDDDDDDDDDDDDDDDDDDDDDDDDDDDDEEEEDEFFFHJIIGIIJJJJIJJJIJJJJJJJIJJIIJJJJJIJHGGJJJJJJJJJJJJHHHHHFFFDD=1#	AS:i:190	XN:i:0	XM:i:0	XO:i:0	XG:i:0	NM:i:0	MD:Z:100	YT:Z:UU
+HWI-ST1410:82:C2VAGACXX:7:1101:11037:2144	16	chrom_1	3637	255	100M1S	*	0	0	TCTCCATGTCGGTTTGAAAGGTGGTGTTAGAAGTCGCATAATCAATCAGGGCGTAAAGTCATACGTGAATTCAGCGGCGAGCACGTTTACATCGAGCGGGT	9DDDDDDDDDDDDDDDDDDDDDDDDCEDDDDDDDDDDDEDEEEEEEDFFHHHJJJJJJIJJIHIJJJJJJJJJJIJJJJJIIHJJJJJHHHHGFFFFFCCB	AS:i:210	XN:i:0	XM:i:0	XO:i:0	XG:i:0	NM:i:0	MD:Z:100	YT:Z:UU
+HWI-ST1410:82:C2VAGACXX:7:1101:14679:2757	0	mito	18	255	92M9S	*	0	0	GTTCTTACGTTACCGTGATTGCAATCTTTTTCCGCCAAGAGTTTAAGAAACTCAAGATTGAAATCCTTTGCGAAGAGTTTAAGAAACTCACCTCCTGAAGT	CCCFFFFFHHHHHJJFHHIJIIJJIIIJJIJJJHHIIIIJJHIIJJJIJJDHIJJIJJIJIJJJJIJGHGHHFFDAC@DDECDDDDDCDDDDDDDDDDDD#	AS:i:153	XN:i:0	XM:i:4	XO:i:0	XG:i:0	NM:i:4	MD:Z:30C1T1T0G56	YT:Z:UU
