--- utils/expectation_maximization/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-    name = "expectation_maximization"
-    version = "0.1.0"
-    edition = "2021"
-    authors = ["Markus Swoveland"]
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
