--- utils/expectation_maximization/TestFiles/rewrite.sam
@@ -1,10 +1,0 @@
-HWI-ST1410:82:C2VAGACXX:7:1101:20066:1892	16	NC_016509	2346	255	100M1S	*	0	0	ACAAAGATTGCAGAAGGCGCGAAGAGGAGTGTTTGAGGCTCGCTACAGCTGAGGAAAAGGTTATGACAGTCGAAAGCGTGCTTAAGAAAGAGGCTGCGAAN	CDDDDDDDDDDDDDDDDDDDDDDDDDDDDEEEEDEFFFHJIIGIIJJJJIJJJIJJJJJJJIJJIIJJJJJIJHGGJJJJJJJJJJJJHHHHHFFFDD=1#	AS:i:200	XN:i:0	XM:i:0	XO:i:0	XG:i:0	NM:i:0	MD:Z:100	YT:Z:UU
-HWI-ST1410:82:C2VAGACXX:7:1101:11037:2144	16	NC_016509	3637	255	100M1S	*	0	0	TCTCCATGTCGGTTTGAAAGGTGGTGTTAGAAGTCGCATAATCAATCAGGGCGTAAAGTCATACGTGAATTCAGCGGCGAGCACGTTTACATCGAGCGGGT	9DDDDDDDDDDDDDDDDDDDDDDDDCEDDDDDDDDDDDEDEEEEEEDFFHHHJJJJJJIJJIHIJJJJJJJJJJIJJJJJIIHJJJJJHHHHGFFFFFCCB	AS:i:200	XN:i:0	XM:i:0	XO:i:0	XG:i:0	NM:i:0	MD:Z:100	YT:Z:UU
-HWI-ST1410:82:C2VAGACXX:7:1101:14679:2757	0	NC_003615	18	255	92M9S	*	0	0	GTTCTTACGTTACCGTGATTGCAATCTTTTTCCGCCAAGAGTTTAAGAAACTCAAGATTGAAATCCTTTGCGAAGAGTTTAAGAAACTCACCTCCTGAAGT	CCCFFFFFHHHHHJJFHHIJIIJJIIIJJIJJJHHIIIIJJHIIJJJIJJDHIJJIJJIJIJJJJIJGHGHHFFDAC@DDECDDDDDCDDDDDDDDDDDD#	AS:i:153	XN:i:0	XM:i:4	XO:i:0	XG:i:0	NM:i:4	MD:Z:30C1T1T0G56	YT:Z:UU
-HWI-ST1410:82:C2VAGACXX:7:1101:15869:3900	16	NC_016509	619	255	101M	*	0	0	TAATTTTTCCTTTTCCTCCTTAGTATTTTGCAGTATCGGTTTCATTTCTAACTTTTTTATTCGTCTTGTACTTCCTTTTATCTTCGCTTTTCTCGTTTTTG	DDDDDDDDDDDDDDDDDDDDDDDDDDDDDDEDDDDDDEEFFFFFFFHEFHAHJIJIJJJJJIIIJJJIHFJJJIJJJJJIGHHGJJJJHGHHHFFFFFCCC	AS:i:202	XN:i:0	XM:i:0	XO:i:0	XG:i:0	NM:i:0	MD:Z:101	YT:Z:UU
-HWI-ST1410:82:C2VAGACXX:7:1101:8981:4601	16	NC_016509	18530	255	101M	*	0	0	GGGAGTGAAATATCGAAATGTTACTTAGTAATAAGATAATCTGAGGGTCTAGGTTCTCAGATTATATAAGTCCTATTAGGGTGAGATATATAAATAAACGG	DDEEEEEFFDFFFHHHHHHHJJIJIHJJJJJIJJJJJJJJJJJJJJJIJJJJJJJJJJJJJIJIIJJJJJJJJJJIJJJJJJJJJJJJHHHHHFFFFFCCC	AS:i:202	XN:i:0	XM:i:0	XO:i:0	XG:i:0	NM:i:0	MD:Z:101	YT:Z:UU
-HWI-ST1410:82:C2VAGACXX:7:1101:3847:4957	0	NC_016509	725	255	101M	*	0	0	CCCCATTGTTCTCTTTTCTTATTTTTATATTTCCTTTCTTCTCGTTTATTTAAAATCATGGCAGTTAAGTCCATCTTTGCACAAGAAGACTTGATTTCCAT	CCCFFFFFFHHHHJJJJJJJJJJJJJJJIJJJJJJJJJJJIJJIHIJJJJJJIIJIIJJJJJJJIJJJJHIJJJJIIJJJJJHHHHHFFFFFEEEEEEEEE	AS:i:202	XN:i:0	XM:i:0	XO:i:0	XG:i:0	NM:i:0	MD:Z:101	YT:Z:UU
-HWI-ST1410:82:C2VAGACXX:7:1101:9319:5506	16	NC_003615	2407	255	100M1S	*	0	0	TAGAAGAAGGCCATAAGATACTGGTTACTGCGGCGGGCGTTCCAAGGAAAACGTCTGCCGATTTTACAATGTGCATCAAGGAAGAGGTATCCAAGTTGGAG	CCCDDDDDDDDDDDEEEDC>BCCC>A?DDDDDFFFHHJJJJJJJJIIJJIHJJIJJJJJIJJJJJIJJJJJJJJJJJJJJJJIJJJJJHHHHHFFFFFCCC	AS:i:141	XN:i:0	XM:i:8	XO:i:0	XG:i:0	NM:i:8	MD:Z:13C11C8A2T20T26A2G2T8	YT:Z:UU
-HWI-ST1410:82:C2VAGACXX:7:1101:11319:6698	16	NC_016509	18528	255	101M	*	0	0	CGGGGAGTGAAATATCGAAATGTTACTTAGTAATAAGATAATCTGAGGGTCTAGGTTCTCAGATTATATAAGTCCTATTAGGGTGAGATATATAAATAAAC	DDDEEEEEFFFFFHHHHHHJJJIHHGJIJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJIJJJJJJJJIJJJJJJHHHHHFFFFFCBB	AS:i:202	XN:i:0	XM:i:0	XO:i:0	XG:i:0	NM:i:0	MD:Z:101	YT:Z:UU
-HWI-ST1410:82:C2VAGACXX:7:1101:17192:6564	0	NC_016509	18473	255	101M	*	0	0	GTTAAATTAAAATAAAATAAAATAAAATAAATAAATGGCTCCTAACGCTAATTACCGGGGAGTGAAATATCGAAATGTTACTTAGTAATAAGATAATCTGA	CCCFFFFFHHHHHJJJJJJJJJJJJJJJJJJJJJIIJJJJJJIJJJJJJJJHJJJJJJJJHHDDFDFEEFEDCDDBCDDEDDDEDDADDFDDDDDDDDDE:	AS:i:202	XN:i:0	XM:i:0	XO:i:0	XG:i:0	NM:i:0	MD:Z:101	YT:Z:UU
-HWI-ST1410:82:C2VAGACXX:7:1101:10743:7728	16	NC_007448	16317	255	2S93M6S	*	0	0	TAAAATCGAACGTGGGTGGATCTACCTGACGTATCAACTTTAGCTGTTACTGGGTAATTAAGCTAACAAGTGTTAGCGTAATGTGTATGTTGATGAAGAGA	@DDDFCDDDDDDDDEEFFFFFFHEHEHJIJJJJJIIHJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJIJJJJJJJJJJJJIJJJJJHHHHHFFFFFCCC	AS:i:123	XN:i:0	XM:i:8	XO:i:0	XG:i:0	NM:i:8	MD:Z:16T21A11A8A1C9C0G1T18	YT:Z:UU
