--- utils/expectation_maximization/TestFiles/to_isolates.sam
@@ -1,11 +1,0 @@
-# This line should be discarded.
-@THIS line should also be discarded.
-HWI-ST1410:82:C2VAGACXX:7:1101:20066:1892	16	NC_016509	2346	255	100M1S	*	0	0	ACAAAGATTGCAGAAGGCGCGAAGAGGAGTGTTTGAGGCTCGCTACAGCTGAGGAAAAGGTTATGACAGTCGAAAGCGTGCTTAAGAAAGAGGCTGCGAAN	CDDDDDDDDDDDDDDDDDDDDDDDDDDDDEEEEDEFFFHJIIGIIJJJJIJJJIJJJJJJJIJJIIJJJJJIJHGGJJJJJJJJJJJJHHHHHFFFDD=1#	AS:i:200	XN:i:0	XM:i:0	XO:i:0	XG:i:0	NM:i:0	MD:Z:100	YT:Z:UU
-HWI-ST1410:82:C2VAGACXX:7:1101:11037:2144	16	NC_016509	3637	255	100M1S	*	0	0	TCTCCATGTCGGTTTGAAAGGTGGTGTTAGAAGTCGCATAATCAATCAGGGCGTAAAGTCATACGTGAATTCAGCGGCGAGCACGTTTACATCGAGCGGGT	9DDDDDDDDDDDDDDDDDDDDDDDDCEDDDDDDDDDDDEDEEEEEEDFFHHHJJJJJJIJJIHIJJJJJJJJJJIJJJJJIIHJJJJJHHHHGFFFFFCCB	AS:i:200	XN:i:0	XM:i:0	XO:i:0	XG:i:0	NM:i:0	MD:Z:100	YT:Z:UU
-HWI-ST1410:82:C2VAGACXX:7:1101:14679:2757	0	NC_003615	18	255	92M9S	*	0	0	GTTCTTACGTTACCGTGATTGCAATCTTTTTCCGCCAAGAGTTTAAGAAACTCAAGATTGAAATCCTTTGCGAAGAGTTTAAGAAACTCACCTCCTGAAGT	CCCFFFFFHHHHHJJFHHIJIIJJIIIJJIJJJHHIIIIJJHIIJJJIJJDHIJJIJJIJIJJJJIJGHGHHFFDAC@DDECDDDDDCDDDDDDDDDDDD#	AS:i:153	XN:i:0	XM:i:4	XO:i:0	XG:i:0	NM:i:4	MD:Z:30C1T1T0G56	YT:Z:UU
-HWI-ST1410:82:C2VAGACXX:7:1101:10743:7728	16	NC_007448	16317	255	2S93M6S	*	0	0	TAAAATCGAACGTGGGTGGATCTACCTGACGTATCAACTTTAGCTGTTACTGGGTAATTAAGCTAACAAGTGTTAGCGTAATGTGTATGTTGATGAAGAGA	@DDDFCDDDDDDDDEEFFFFFFHEHEHJIJJJJJIIHJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJIJJJJJJJJJJJJIJJJJJHHHHHFFFFFCCC	AS:i:123	XN:i:0	XM:i:8	XO:i:0	XG:i:0	NM:i:8	MD:Z:16T21A11A8A1C9C0G1T18	YT:Z:UU
-HWI-ST1410:82:C2VAGACXX:7:1101:3175:8662	256	NC_017939	346	9	101M	*	0	0	CGGAGGCCAAGGAACTCCGCTACGTGTGCTCATGTTGGATGAGCACACGTCTCGTTAAAGTTGAGGCACCTCCTCAACAGTCAAGGAAGAGTGGGATAGCG	@C@FFFFFGHGHHJJIJJIIIJJIIG?DGADA?DBF>G9@B9B?DFCC<8CC=C;@=?E;7?>CBD7=ACCBDD?@@9:9,:>A:A??8??+3<<24<AC9	AS:i:143	XS:i:143	XN:i:0	XM:i:9	XO:i:0	XG:i:0	NM:i:9	MD:Z:7T42T1A5G1C6T8G2A7G13	YT:Z:UU
-HWI-ST1410:82:C2VAGACXX:7:1101:20350:9357	0	NC_003623	1182	255	100M1S	*	0	0	CTTTTGAGCAGGCCAGGAATGCTTATCTTGGTGCTGCTGATATGATTGAGCAAGGCCTGCCGCTGCTTCCCCCTCTGCGCAGCGCTTATGCCCCTAGGGGC	CCCFFFFFHHHHHJIIJIJJJJJJJJJJJJJGIJIJJJJIJJJIIJJIIJJJJIJIIIJJJJJJHHHHFFFFDDDDDDDDDDDDDBDDDDDDDDDDDDDDB	AS:i:169	XN:i:0	XM:i:4	XO:i:0	XG:i:0	NM:i:4	MD:Z:37C2C17A29C11	YT:Z:UU
-HWI-ST1410:82:C2VAGACXX:7:1101:5156:10191	16	NC_016509	16954	255	101M	*	0	0	TGCTCCAGTGTGTTATATTTGTACTAAGAATGAAGAAGGAAGTGATAAGAAATTAACGCTTTTGTCTCGTGAGGATCTCATCGATTTTAACAGTGACGAGA	DDDCDDDDDCDCDEEEEEEEEEFEFEFFFHHHHHHHJJHHIIGIIIJJJJJJJJHGJIJJJJJIJJHJJIJJJJJJJIIJJJJJJJJJHHHHHFFFFFCCC	AS:i:202	XN:i:0	XM:i:0	XO:i:0	XG:i:0	NM:i:0	MD:Z:101	YT:Z:UU
-HWI-ST1410:82:C2VAGACXX:7:1101:20113:10836	0	NC_003615	3403	255	101M	*	0	0	ATGTAATCCCATGGTTAGAAAAGAAAACATTGGAAATTGTACATTGGGATGTACATAGACATATTGCCACTGGTCCTCGTAATGCACTGGTTGCATGCTTT	CCCFFFFFHHHHHJHIIJJJJJJJJJJJJJJJJJHHHIIHIJJJJJJJIIJIGIJJJJIJJJIJIIJJIJIJJIJHIIHHGFFFFFFEEEEEDDDEDDDED	AS:i:142	XN:i:0	XM:i:8	XO:i:0	XG:i:0	NM:i:8	MD:Z:4G2T8G14A20G4A18C2C21	YT:Z:UU
-HWI-ST1410:82:C2VAGACXX:7:1101:20066:1892	16	NC_003347	6551	255	9S92M	*	0	0	TTCCGATCTCTCGAGTAAGCATCCACCTCCCCTTCCAGTTCCTGTGGTATGACATCACAGGCACCGAGTCCTCCTACACCTCCCTGTCCATCGCCTCCCGC	BBDDDDDDDDDDDDDDEDDDDDBA8@DBDDDDDDDDDDDDDDDEEECEFFFEFFFFHHHHHHEJJJJIHJJIJJHFF?JHHJJJJJJJHHHHHFFFFFCCC	AS:i:189	XN:i:0	XM:i:3	XO:i:0	XG:i:0	NM:i:3	MD:Z:4T11T32G42	YT:Z:UU
